add fix to make the properties list responsive
